Got rid of redundant event emitters

+++ src/app/wiki/wiki.component.ts
@@ -3,7 +3,7 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { WikiModel } from '../_shared/models';
 import { WikiService } from './wiki.service';
 import { take } from 'rxjs/operators';
-import { Observable, Subscription } from 'rxjs';
+import { Observable, Subscription, VirtualTimeScheduler } from 'rxjs';
 import { NotificationService } from '../_shared/notification.service';
 
 @Component({
@@ -46,37 +46,49 @@ export class WikiComponent implements OnInit, OnDestroy {
   }
 
   savePageEdits(page) {
-    this.wikiService.savePageChanges(page)
-      .pipe(take(1))
-      .subscribe(() => {
-        this.notificationService.notify('success', 'Page updated!');
-        this.toggleEditForm();
-        this.getPageContent(page.path);
-      }, () => {
-        this.notificationService.notify('error', 'Unable to update page');
-      })
+    if (page) {
+      this.wikiService.savePageChanges(page)
+        .pipe(take(1))
+        .subscribe(() => {
+          this.notificationService.notify('success', 'Page updated!');
+          this.toggleEditForm();
+          this.getPageContent(page.path);
+        }, () => {
+          this.notificationService.notify('error', 'Unable to update page');
+        })
+    } else {
+      this.toggleEditForm();
+    }
   }
 
   addNewPage(path) {
-    this.wikiService.createPage(path)
-    .subscribe(() => {
-      this.notificationService.notify('success', 'New page created!');
+    if (path) {
+      this.wikiService.createPage(path)
+      .subscribe(() => {
+        this.notificationService.notify('success', 'New page created!');
+        this.toggleAddForm();
+        this.router.navigate(['/wiki/', path]);
+      }, () => {
+        this.notificationService.notify('error', 'Unable to create new page');
+      });
+    } else {
       this.toggleAddForm();
-      this.router.navigate(['/wiki/', path]);
-    }, () => {
-      this.notificationService.notify('error', 'Unable to create new page');
-    });
+    }
   }
 
   deletePage(path) {
-    this.wikiService.deletePage(path)
-    .subscribe(() => {
-      this.notificationService.notify('success', 'Page deleted!');
+    if (path) {
+      this.wikiService.deletePage(path)
+      .subscribe(() => {
+        this.notificationService.notify('success', 'Page deleted!');
+        this.toggleDeleteConfirm();
+        this.router.navigateByUrl('/');
+      }, () => {
+        this.notificationService.notify('error', 'Unable to delete page');
+      });
+    } else {
       this.toggleDeleteConfirm();
-      this.router.navigateByUrl('/');
-    }, () => {
-      this.notificationService.notify('error', 'Unable to delete page');
-    });
+    }
   }
 
   toggleEditForm() {
